refactor: extract types and enums
