+++ packages/ckbtc/src/enums/btc.enums.ts
@@ -1,0 +1,11 @@
+export enum BtcNetwork {
+    Mainnet,
+    Regtest,
+    Testnet,
+}
+
+export enum BtcAddressType {
+    P2wpkhV0,
+    P2pkh,
+    P2sh,
+}
