+++ packages/ckbtc/src/index.ts
@@ -3,8 +3,10 @@ export type {
   RetrieveBtcOk,
   UpdateBalanceResult,
 } from "../candid/minter";
+export * from "./enums/btc.enums";
 export * from "./errors/minter.errors";
 export { CkBTCMinterCanister } from "./minter.canister";
+export * from "./types/btc";
 export * from "./types/minter.params";
 export * from "./types/minter.responses";
 export * from "./utils/btc.utils";
