+++ packages/ckbtc/src/types/btc.ts
@@ -1,0 +1,6 @@
+import type {BtcNetwork} from "../enums/btc.enums";
+
+export interface BtcAddress {
+  address: string;
+  network?: BtcNetwork;
+}
